Enemy Movement Update
Added speed property for Enemy objects along with a minimum and
maximum speed. The speed property is used to define the amount of
movement each enemy object makes for each game tick.

When enemy objects leave the screen, they are redrawn to the left of the
canvas and assigned a new random row.

Signed-off-by: Avery Uslaner <uslaner.avery@gmail.com>